# src/main/java/com/example/bankcards/dto/requests/BalanceRequest.java
package com.example.bankcards.dto.requests;

import com.fasterxml.jackson.annotation.JsonProperty;
import jakarta.validation.constraints.NotNull;
import lombok.*;
import lombok.experimental.FieldDefaults;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@FieldDefaults(level = AccessLevel.PRIVATE)
public class BalanceRequest {
    @JsonProperty("last_four_card_digits")
    @NotNull
    List<String> lastFourCardDigits;
}